requirements.txt
fastapi
uvicorn[standard]
qrcode[pil]
pydantic
PyJWT
cryptography